Add methods for indexing plus start tree-type
New api methods PUT and POST for populating modules
and vendors. This will provide data for indexing and
it will be either all data need to be reindexed or only
the one that are not currently in yang catalog.

Start new metadata parsing - tree type - this is not
done yet.

Cleaning up code a little.

+++ tools/api/api.py
@@ -11,6 +11,7 @@ import subprocess
 import unicodedata
 import ConfigParser
 import urllib2
+from urllib2 import URLError
 
 import MySQLdb
 import datetime
@@ -57,6 +58,8 @@ def http_request(path, method, json_data, http_credentials, header):
         if method == 'DELETE':
             return
         print('Could not send request with body ' + repr(json_data) + ' and path ' + path)
+        raise e
+    except URLError as e:
         raise e
 
 
@@ -215,24 +218,31 @@ def delete_modules(name, revision):
     return jsonify({'info': 'success'})
 
 
-
-@app.route('/modules', methods=['PUT'])
+@app.route('/modules', methods=['PUT', 'POST'])
 @auth.login_required
 def add_modules():
     if not request.json:
         abort(400)
     body = request.json
- #   http_request(protocol + '://' + confd_ip + ':' + repr(confdPort) + '/api/config/modules', 'PUT', json.dumps(body),
- #                credentials)
+    tree_created = False
+    #   http_request(protocol + '://' + confd_ip + ':' + repr(confdPort) + '/api/config/modules', 'PUT', json.dumps(body),
+    #                credentials)
     with open('./prepare-sdo.json', "w") as plat:
-         json.dump(body, plat)
+        json.dump(body, plat)
 
     repo = {}
     warning = []
     for mod in body['modules']['module']:
         sdo = mod['source-file']
         orgz = mod['organization']
-
+        if request.method == 'POST':
+            try:
+                path = protocol + '://' + confd_ip + ':' + repr(confdPort) + '/api/config/catalog/modules/module/' + \
+                       mod['name'] + ',' + mod['revision']
+                http_request(path, 'GET', None, credentials, 'application/vnd.yang.data+json')
+                continue
+            except urllib2.HTTPError as e:
+                pass
         directory = '/'.join(sdo['path'].split('/')[:-1])
 
         repo_url = url + sdo['owner'] + '/' + sdo['repository']
@@ -251,8 +261,7 @@ def add_modules():
             if e.errno != errno.EEXIST:
                 raise
         shutil.copy(repo[repo_url].localdir + '/' + sdo['path'], save_to)
-        #organization = yangParser.parse(os.path.abspath(save_to + '/' + sdo['path'].split('/')[-1])) \
-        #    .search('organization')[0].arg
+        tree_created = True
         organization = ''
         try:
             namespace = yangParser.parse(os.path.abspath(save_to + '/' + sdo['path'].split('/')[-1])) \
@@ -295,12 +304,15 @@ def add_modules():
     with open("log.txt", "wr") as f:
         try:
             arguments = ["python", "../parseAndPopulate/populate.py", "--sdo", "--port", repr(confdPort), "--dir",
-                          "temp", "--api", "--ip", confd_ip, "--credentials", credentials[0], credentials[1]]
+                         "temp", "--api", "--ip", confd_ip, "--credentials", credentials[0], credentials[1]]
             subprocess.check_call(arguments, stderr=f)
         except subprocess.CalledProcessError as e:
             shutil.rmtree('temp/')
             for key in repo:
                 repo[key].remove()
+            for item in os.listdir('./'):
+                if item.endswith(".json"):
+                    os.remove(item)
             return jsonify({'error': 'Was not able to write all or partial yang files'})
 
     try:
@@ -309,31 +321,54 @@ def add_modules():
         # be happy if someone already created the path
         if e.errno != errno.EEXIST:
             raise
-    subprocess.call(["cp", "-r", "temp/.", "../../api/sdo/"])
 
-    shutil.rmtree('temp')
+    if tree_created:
+        subprocess.call(["cp", "-r", "temp/.", "../../api/sdo/"])
+        shutil.rmtree('temp')
+        send_to_indexing('./prepare-sdo.json')
+
     for item in os.listdir('./'):
-        if 'log' in item and '.txt' in item:
+        if item.endswith(".json"):
             os.remove(item)
     for key in repo:
         repo[key].remove()
     if len(warning) > 0:
-        return jsonify({'info': 'success', 'warnings': [{'warning': val}for val in warning]})
+        return jsonify({'info': 'success', 'warnings': [{'warning': val} for val in warning]})
     else:
         return jsonify({'info': 'success'})
 
 
-@app.route('/platforms', methods=['PUT'])
+def send_to_indexing(file_to_index):
+    with open(file_to_index, 'r') as f:
+        sdos_json = json.load(f)
+    post_body = {}
+    for sdo in sdos_json['module']:
+        post_body[sdo['name'] + '@' + sdo['revision']] = sdo['source-file']['local']['path']
+    body_to_send = json.dumps({'modules-to-index': post_body})
+    try:
+        http_request('https://' + confd_ip + '/yang-search/metadata-update.php', 'POST', body_to_send,
+                     credentials, 'application/json')
+    except urllib2.HTTPError as e:
+        print('could not send data for indexing. Reason: ' + e.msg)
+    except URLError as e:
+        print('could not send data for indexing. Reason: ' + repr(e.message))
+
+
+@app.route('/platforms', methods=['PUT', 'POST'])
 @auth.login_required
 def add_vendors():
     if not request.json:
         abort(400)
     body = request.json
+    tree_created = False
     resolved_authorization = authorize_for_vendors(request, body)
     if 'passed' != resolved_authorization:
         return resolved_authorization
-    http_request(protocol + '://' + confd_ip + ':' + repr(confdPort) + '/api/config/platforms/',
-                 'DELETE', None, credentials, 'application/vnd.yang.collection+json')
+    try:
+        http_request(protocol + '://' + confd_ip + ':' + repr(confdPort) + '/api/config/platforms/',
+                     'DELETE', None, credentials, 'application/vnd.yang.collection+json')
+    except:
+        pass
     path = protocol + '://' + confd_ip + ':' + repr(confdPort) + '/api/config/'
     try:
         http_request(path, 'POST', json.dumps(body), credentials, 'application/vnd.yang.data+json')
@@ -343,8 +378,13 @@ def add_vendors():
     for platform in body['platforms']:
         capability = platform['capabilities-file']
         file_name = capability['path'].split('/')[-1]
+        if request.method == 'POST':
+            repo_split = capability['repository'].split('.')[0]
+            if os.path.isfile('../../' + capability['owner'] + '/' + repo_split + '/' + capability['path']):
+                continue
 
         repo_url = url + capability['owner'] + '/' + capability['repository']
+
         if repo_url not in repo:
             repo[repo_url] = repoutil.RepoUtil(repo_url)
             repo[repo_url].clone()
@@ -363,6 +403,7 @@ def add_vendors():
         with open(save_to + '/' + file_name.split('.')[0] + '.json', "w") as plat:
             json.dump(platform, plat)
         shutil.copy(repo[repo_url].localdir + '/' + capability['path'], save_to,)
+        tree_created = True
 
     with open("log.txt", "wr") as f:
         try:
@@ -380,11 +421,15 @@ def add_vendors():
         # be happy if someone already created the path
         if e.errno != errno.EEXIST:
             raise
+
     subprocess.call(["cp", "-r", "temp/.", "../../api/vendor/"])
 
-    shutil.rmtree('temp/')
+    if tree_created:
+        shutil.rmtree('temp/')
+        send_to_indexing('./prepare.json')
+
     for item in os.listdir('./'):
-        if 'log' in item and '.txt' in item:
+        if item.endswith(".json"):
             os.remove(item)
     for key in repo:
         repo[key].remove()
@@ -424,6 +469,14 @@ def search(key, value):
                 }), mimetype='application/json')
             else:
                 return Response(mimetype='application/json', status=204)
+
+
+@app.route('/search/modules/<name>,<revision>', methods=['GET'])
+def search_module(name, revision):
+    path = protocol + '://' + confd_ip + ':' + repr(confdPort) + '/api/config/catalog/modules/module/' + name + ','\
+           + revision + '?deep'
+    data = json.loads(http_request(path, 'GET', '', credentials, 'application/vnd.yang.data+json').read())
+    return Response(json.dumps(data), mimetype='application/json')
 
 
 def process(data, passed_data, value, module, split, count):
